Stores & combines every levelScene encountered as a map of the whole world.
This map is made up of MapSquare classes, which currently only store the encoding they're representing but could store all other MapSquares it's possible for Mario to move into from the current one. This may make route-planning easier.
This change allows a route to get a coin/kill an enemy possible, as we have a fuller picture of the environment.
TODO: swap all uses of levelScene to map as they now store the same data about the immediate environment so no need to keep levelScene. Need to consider how to deal with meeting enemies on a planned route before this is done, as it may require the retention of levelScene for finding enemies & not storing them in map (or just ignoring enemies when route-finding).