Added MapSquareWrapper class & modified A* to use it.
Changed the A* method to use MapSquareWrapper instead of MapSquare - this allows multiple values of g,h to be stored for each MapSquare (& therefore means A* can actually work). Note: need to implement stopping it from repeatedly checking its parent nodes in the search tree (currently works because going in a loop will never be the fastest way, but if the node is unreachable current implementation won't detect that & will loop forever).
Added encapsulation for adjacent nodes when checking reachable squares in MapSquare.
Fixed location bug with MapUpdater (all MapSquares were previously storing their x,y values as reversed).
Changed SQUARESIZE so hopefully the map will now be an accurate representation of the world.