Added test for A* which runs the A* algorithm on a hard-coded map
Changed how reachable squares reached by jumping are detected. Now the above square is only included by workOutReachableSquares(..) when the current square has a floor to jump off. If there is no floor, when reachable squares are returned, it requires whether you are currently jumping & uses this to decide whether the square above can be reached (based on its contents & your current stage in the jump). Note: this should also be done with moving left/right (as can only do it if there's a floor, or we entered this square from above - cannot move right/left indefinitely with no ground).
Mended bug with workOutReachableSquares(..) where it was calculating them while creating MapSquares, so the first square never had any reachable squares as none had been created yet. This involved moving the method into MapUpdater, so it was sure to be called everytime the map was updated. Note: this can be optimised by storing which MapSquares are new or have had their value changed & updating them & their neighbours.