Started using a plan to decide the next move - this meant changing the backtracked route to use a Stack for ease of transfer (note: may want to change this, as sometimes don't use it exactly like a Stack)
Changed how blockages are recorded - getBlockageLocation(..) now returns the first empty square above the blockage, which allows a plan to be made to get there
Added methods to MapUpdater which translate levelScene coordinates into map coordinates (and vice versa)
Added cut off point to the A* - if we have travelled more than 15 squares to get to this & are not closer than 5 to the target, we ignore this square. This significantly reduces the time taken, especially when the target is unreachable.