Changed from xml for saving to serialising the list of LSMs (and changed filename to [oldname].ser)
Changed fitness function to only consider amount moved right in the last frame (as opposed to the total movement right).
Added ActionsIndex which stores all key states that don't have conflicting instructions (e.g. pressing both left and right) and can translate from an int to the corresponding array and vice versa
Added boolean array to LSM which stores which "actions" have been tried, so we only try ones that haven't previously been attempted (and were worse than the current set of actions for that levelscene)
Changed when a new action array is chosen for the current LevelScene - now mutates according to evolver.getSizeOfGeneration() instead of proportional to reward