Added some simplifications to the level scene stored by a LSM to improve genetic algorithm
Reduced size of LevelScene stored by LSM, as Mario can only interact with the nearest 5 squares so anything further away than that is irrelevant - this reduces the amount of work needing to be done by the genetic algorithm as it needs to match fewer numbers to get an exact match.
Added method to Encoding to allow for simplifying of the Encodings of LevelScene (maps all Environment pieces to the same byte, as Mario doesn't need to discern between them).
Now always update a stored LSM with the most recent reward for a set of actions, and replace an equal rewarded set of actions with a probability of 0.5.
Fixed equals bug when comparing boolean arrays.